+++ lib/oli/institutions/institution.ex
@@ -12,6 +12,7 @@ defmodule Oli.Institutions.Institution do
     # LTI 1.3 Deployments
     has_many :registrations, Oli.Lti_1p3.Tool.Registration
     has_many :sections, Oli.Delivery.Sections.Section
+    has_many :brands, Oli.Branding.Brand
 
     timestamps(type: :utc_datetime)
   end