+++ lib/oli_web/plugs/require_section.ex
@@ -2,12 +2,14 @@ defmodule Oli.Plugs.RequireSection do
   import Plug.Conn
   import Phoenix.Controller
 
+  alias Oli.Delivery.Sections
+
   def init(opts), do: opts
 
   def call(conn, _opts) do
     case conn.path_params do
       %{"section_slug" => section_slug} ->
-        case Oli.Delivery.Sections.get_section_by(slug: section_slug) do
+        case Sections.get_section_by_slug(section_slug) do
           nil ->
             section_not_found(conn)