+++ lib/oli_web/views/brand_view.ex
@@ -1,0 +1,3 @@
+defmodule OliWeb.BrandView do
+  use OliWeb, :view
+end